♻️  Further split mempool in verify_tx / block_construction, rename new_tx (#871)
